+++ pkg/prometheus/resource_selector_test.go
@@ -2646,8 +2646,7 @@ func TestSelectScrapeConfigs(t *testing.T) {
 					},
 				}
 			},
-			promVersion: "3.7.0",
-			valid:       true,
+			valid: true,
 		},
 		{
 			scenario: "EC2 SD config with no secret ref provided",
@@ -2658,8 +2657,7 @@ func TestSelectScrapeConfigs(t *testing.T) {
 					},
 				}
 			},
-			promVersion: "3.7.0",
-			valid:       true,
+			valid: true,
 		},
 		{
 			scenario: "EC2 SD config with invalid secret ref for secretKey",
@@ -2682,8 +2680,7 @@ func TestSelectScrapeConfigs(t *testing.T) {
 					},
 				}
 			},
-			promVersion: "3.7.0",
-			valid:       false,
+			valid: false,
 		},
 		{
 			scenario: "EC2 SD config with valid TLS Config",
@@ -2718,8 +2715,7 @@ func TestSelectScrapeConfigs(t *testing.T) {
 					},
 				}
 			},
-			promVersion: "3.7.0",
-			valid:       true,
+			valid: true,
 		},
 		{
 			scenario: "EC2 SD config with valid HTTPS Config",
@@ -2778,8 +2774,7 @@ func TestSelectScrapeConfigs(t *testing.T) {
 					},
 				}
 			},
-			promVersion: "3.7.0",
-			valid:       false,
+			valid: false,
 		},
 		{
 			scenario: "EC2 SD config with valid proxy settings",
@@ -2807,18 +2802,6 @@ func TestSelectScrapeConfigs(t *testing.T) {
 			},
 			promVersion: "2.52.0",
 			valid:       true,
-		},
-		{
-			scenario: "EC2 SD config with unsupported version",
-			updateSpec: func(sc *monitoringv1alpha1.ScrapeConfigSpec) {
-				sc.EC2SDConfigs = []monitoringv1alpha1.EC2SDConfig{
-					{
-						Region: ptr.To("us-east-1"),
-					},
-				}
-			},
-			promVersion: "3.8.0",
-			valid:       false,
 		},
 		{
 			scenario: "Azure SD config with valid options for OAuth authentication method",
@@ -4057,8 +4040,7 @@ func TestSelectScrapeConfigs(t *testing.T) {
 					},
 				}
 			},
-			promVersion: "3.7.0",
-			valid:       true,
+			valid: true,
 		},
 		{
 			scenario: "LightSail SD config with invalid TLS config with invalid CA data",
@@ -4104,8 +4086,7 @@ func TestSelectScrapeConfigs(t *testing.T) {
 					},
 				}
 			},
-			promVersion: "3.7.0",
-			valid:       true,
+			valid: true,
 		},
 		{
 			scenario: "LightSail SD config with invalid proxy settings",
@@ -4129,8 +4110,7 @@ func TestSelectScrapeConfigs(t *testing.T) {
 					},
 				}
 			},
-			promVersion: "3.7.0",
-			valid:       false,
+			valid: false,
 		},
 		{
 			scenario: "LightSail SD config with invalid secret ref",
@@ -4148,8 +4128,7 @@ func TestSelectScrapeConfigs(t *testing.T) {
 					},
 				}
 			},
-			promVersion: "3.7.0",
-			valid:       false,
+			valid: false,
 		},
 
 		{
@@ -4173,8 +4152,7 @@ func TestSelectScrapeConfigs(t *testing.T) {
 					},
 				}
 			},
-			promVersion: "3.7.0",
-			valid:       true,
+			valid: true,
 		},
 		{
 			scenario: "LightSail SD config with no secret ref provided",
@@ -4185,8 +4163,7 @@ func TestSelectScrapeConfigs(t *testing.T) {
 					},
 				}
 			},
-			promVersion: "3.7.0",
-			valid:       true,
+			valid: true,
 		},
 		{
 			scenario: "LightSail SD config with invalid secret ref for accessKey",
@@ -4209,8 +4186,7 @@ func TestSelectScrapeConfigs(t *testing.T) {
 					},
 				}
 			},
-			promVersion: "3.7.0",
-			valid:       false,
+			valid: false,
 		},
 		{
 			scenario: "LightSail SD config with invalid secret ref for secretKey",
@@ -4233,20 +4209,7 @@ func TestSelectScrapeConfigs(t *testing.T) {
 					},
 				}
 			},
-			promVersion: "3.7.0",
-			valid:       false,
-		},
-		{
-			scenario: "LightSail SD config with unsupported version",
-			updateSpec: func(sc *monitoringv1alpha1.ScrapeConfigSpec) {
-				sc.LightSailSDConfigs = []monitoringv1alpha1.LightSailSDConfig{
-					{
-						Region: ptr.To("us-east-1"),
-					},
-				}
-			},
-			promVersion: "3.8.0",
-			valid:       false,
+			valid: false,
 		},
 		{
 			scenario: "OVHCloud SD config",